feat: add support for firefox
Replaced all calls to the Chrome API with calls to the
webextension-polyfill.

In Manifest v3, Firefox still uses background.scripts instead of
background.service_worker. To ensure compatibility with both browsers,
we include both entries in the manifest file. The unsupported entry will
be ignored, but may trigger a warning in developer mode.